ARM: dts: qcom: Add IOMMU property for CCU QUP node
For the CCU QUP use case, a dedicated SID is assigned to the
GSI node but not to the QUP wrapper node because the CCU QUP
operates only in GSI mode for the Camera use case.
During transfer, the I2C driver maps the DMA buffer using
the dma_map_single API, passing the wrapper node as the dev
node. The IOMMU driver checks the IOMMU ops associated with
that dev node, but without an IOMMU property assigned, it
performs direct mapping for the buffer. This causes a GSI
bus error when the GSI hardware tries to access the DMA
address, as the mapped address is not linked to the GSI node.

To avoid this issue, assign the same IOMMU property to both
the wrapper QUP node and the GSI node.

Change-Id: I19b62b4f3eb120dffb1a1cdad27d4b5242a9672a
Signed-off-by: Aniket Randive <quic_arandive@quicinc.com>