ARM: dts: qcom: sa8797p-overlay: Add tzcom heap support
The tzcom heap is a secure memory region used for sharing
buffers between the non-secure world (Linux kernel) and
the secure world (TrustZone). This is typically required
for use cases involving secure services accessed via the
SMCINVOKE interface, such as trusted applications
or secure firmware components.

By defining the "qcom,tzcom" heap in the device tree, we
enable user space and kernel components to allocate memory
from this heap for secure communication. This is essential
for platforms that rely on TrustZone-based services for
security-sensitive operations.

Change-Id: I48402388808233245ba153f9022c5fad5d911127
Signed-off-by: Yiftach Ariel <quic_yariel@quicinc.com>