ARM: dts: qcom: sa8797p: Add address-cells and size-cells property to interrupt-controller node
Updated the missing address-cells and size-cells property in the GIC v3
interrupt controller. This ensures proper parsing of interrupts for
child nodes.
During the parsing of the interrupt map for nodes like pcie, interrupt
parent node are checked for the address-cells property. If the interrupt
controller node does not have the address-cells property defined, the
value will be considered as 0, leading to incorrect parsing of the
interrupt map.

Change-Id: I1718ada1a7cb6364f19ece89247641173b4eb2a4
Signed-off-by: Jai Kumar Gautam <jgautam@qti.qualcomm.com>