ARM: dts: qcom: sa8797p: Add overlay dt to configure VM memory node
Since target devices can have varying DDR configurations, the
memory allocation for the VM changes based on the DDR size
available on the target. To accommodate this, we create overlay
device trees for different VM memory node configurations.

ABL selects the appropriate device tree (DT) based on the
subtype value in the board-id. Bits 8-10 in the board-id
subtype field are reserved for DDR type. This subtype value
is updated during the overlay process.

Change-Id: I6087feaa2e42b4504f7fe01dd5130b157c179897
Signed-off-by: Jai Kumar Gautam <jgautam@qti.qualcomm.com>