ARM: dts: qcom: ethernet: Add per-channel DMA interrupts for eth0
The ETHQOS Ethernet controller on the SA8797P platform
supports per-channel interrupts for both transmit (Tx)
and receive (Rx) DMA operations.

This commit updates the device tree to define and assign
the full set of available DMA interrupts (10 Rx and 10 Tx)
to the 'ethqos' node. The'interrupt-names' property is
populated accordingly to allow the stmmac driver to probe
and utilize these interrupts for multi-queue operation.

Change-Id: Ib02133462b3f1b472a3ac49137dfb4895c06d1df
Signed-off-by: Sagar Cheluvegowda <quic_scheluve@quicinc.com>